Improve Resource.add to Resource.until.
Resource.until still paginates through all pages like Resource.all.
However Resource.until has extra functionality. It accepts a callback
function which it fires for every page paginated.

In this callback the resource.data and resource is passed through for
convinience. If the callback returns true then the pagination ceases.

The callback you are able to find and filter for data.